requirements.txt
# Fashion AI Pipeline Requirements

# Core ML/DL libraries
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0

# Data processing
scikit-learn>=1.0.0
pandas>=1.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Web framework
Flask>=2.0.0
Werkzeug>=2.0.0

# COCO API (for evaluation)
pycocotools>=2.0.0

# Progress bars
tqdm>=4.62.0

# Image processing
scikit-image>=0.18.0

# Additional utilities
dataclasses-json>=0.5.0

